// src/app/vendor/model/organization.model.ts
export class Organization {
    organizationId?: number;
    organizationName?: string;
    address?: string;
    address2?: string;
    city?: string;
    state?: string;
    zipCode?: string;
    dateOfIncorporation?: Date;
    organizationTypeId?: number;
    duration?: number;
    taxId?: string;
    phone?: string;
    fax?: string;  
  }